Refactor getRawDataFromFile logic and add deltas from excel support

+++ diploma-backend/src/main/java/ru/merenkov/diploma/domain/FuzzyNumber.java
@@ -1,9 +1,11 @@
 package ru.merenkov.diploma.domain;
 
 
+import java.math.BigDecimal;
+
 public record FuzzyNumber(
-        Double largestValue,
-        Double value,
-        Double smallestValue
+        BigDecimal largestValue,
+        BigDecimal value,
+        BigDecimal smallestValue
 ) {
 }
